+++ src/app/orders/models/order.model.ts
@@ -4,6 +4,9 @@ export class OrderModel {
   constructor(
     public clientFirstName: string,
     public clientLastName: string,
+    public clientEmail: string,
+    public clientPhone: string,
+    public clientAddress: string,
     public items: CartItem[],
     public totalSum: number,
     public id?: number,