+++ src/app/orders/orders.module.ts
@@ -1,17 +1,22 @@
 import { NgModule } from '@angular/core';
+import { ReactiveFormsModule } from '@angular/forms';
 
 import { OrdersRoutingModule } from './orders-routing.module';
 import { SharedModule } from './../shared/shared.module';
 import { LocalStorageService } from '../core/services/local-storage.service';
+import { ClientEmailDirective } from './validators/client-email.directive';
 
 @NgModule({
   declarations: [
-    OrdersRoutingModule.components
+    OrdersRoutingModule.components,
+    ClientEmailDirective
   ],
   imports: [
     SharedModule,
-    OrdersRoutingModule
+    OrdersRoutingModule,
+    ReactiveFormsModule
   ],
+  exports: [ClientEmailDirective],
   providers: [
     { provide: LocalStorageService, useClass: LocalStorageService }
   ]